fix(kitchenpos) : 컨텍스트 별로 나눠 Fixture 생성
