+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -2,7 +2,8 @@ package kitchenpos.application;
 
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuGroupRepository;
-import kitchenpos.fixtures.Fixture;
+import kitchenpos.fixtures.FixtureMenu;
+import kitchenpos.fixtures.FixtureOrder;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -23,7 +24,7 @@ class MenuGroupServiceTest {
   @Test
   @DisplayName("메뉴 그룹을 등록하기 위해 메뉴 그룹 이름 입력해야 한다.")
   void case1() {
-    final MenuGroup menuGroup = Fixture.fixtureMenuGroup();
+    final MenuGroup menuGroup = FixtureMenu.fixtureMenuGroup();
     given(menuGroupRepository.save(any())).willReturn(menuGroup);
 
     final MenuGroup resultMenuGroup = menuGroupService.create(menuGroup);
@@ -34,7 +35,7 @@ class MenuGroupServiceTest {
   @DisplayName("메뉴 그룹을 전체 조회할 수 있다.")
   void case2() {
     given(menuGroupRepository.findAll())
-        .willReturn(List.of(Fixture.fixtureMenuGroup()));
+        .willReturn(List.of(FixtureMenu.fixtureMenuGroup()));
 
     final List<MenuGroup> all = menuGroupService.findAll();
     Assertions.assertThat(1).isEqualTo(all.size());
