+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -4,7 +4,6 @@ import static org.mockito.BDDMockito.*;
 
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -34,9 +33,9 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴를 등록하기 위해 메뉴 그룹, 이름을 입력해야 한다.")
   void case1() {
-    final Menu menu = Fixture.fixtureMenu();
-    final Product product = Fixture.fixtureProduct();
-    final MenuGroup menuGroup = Fixture.fixtureMenuGroup();
+    final Menu menu = FixtureMenu.fixtureMenu();
+    final Product product = FixtureProduct.fixtureProduct();
+    final MenuGroup menuGroup = FixtureMenu.fixtureMenuGroup();
 
     menu.setName("닭다리");
 
@@ -54,9 +53,9 @@ class MenuServiceTest {
   @DisplayName("메뉴를 등록하기 위해 메뉴 이름에 비속어 또는 욕설을 작성할 수 없다.")
   void case2() {
 
-    final Menu menu = Fixture.fixtureMenu();
-    final Product product = Fixture.fixtureProduct();
-    final MenuGroup menuGroup = Fixture.fixtureMenuGroup();
+    final Menu menu = FixtureProduct.fixtureMenu();
+    final Product product = FixtureProduct.fixtureProduct();
+    final MenuGroup menuGroup = FixtureMenu.fixtureMenuGroup();
 
     given(menuGroupRepository.findById(any())).willReturn(Optional.of(menuGroup));
     given(productRepository.findAllByIdIn(any())).willReturn(List.of(product));
@@ -73,7 +72,7 @@ class MenuServiceTest {
   @ValueSource(longs = {-1})
   @DisplayName("메뉴를 등록하기 위해 금액을 입력하지 않았거나 또는 0원 밑으로 입력할 수 없다.")
   void case3(Long value) {
-    final Menu menu = Fixture.fixtureMenu();
+    final Menu menu = FixtureMenu.fixtureMenu();
 
     menu.setName("닭다리");
     menu.setPrice(value == null ? null : BigDecimal.valueOf(value));
@@ -86,10 +85,10 @@ class MenuServiceTest {
   @ValueSource(ints = {-1})
   @DisplayName("메뉴를 등록하기 위해 상품이 있어야하고, 수량을 0개 밑으로 입력할 수 없다.")
   void case4(Integer value) {
-    final Menu menu = Fixture.fixtureMenu();
-    final Product product = Fixture.fixtureProduct();
-    final MenuGroup menuGroup = Fixture.fixtureMenuGroup();
-    final MenuProduct menuProduct = Fixture.fixtureMenuProduct();
+    final Menu menu = FixtureMenu.fixtureMenu();
+    final Product product = FixtureProduct.fixtureProduct();
+    final MenuGroup menuGroup = FixtureMenu.fixtureMenuGroup();
+    final MenuProduct menuProduct = FixtureProduct.fixtureMenuProduct();
 
     menuProduct.setQuantity(value);
     menu.setMenuProducts(List.of(menuProduct));
@@ -104,14 +103,14 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴를 등록하기 위해 해당 상품 정보가 상품에 존재해야 한다.")
   void case5() {
-    final Menu menu = Fixture.fixtureMenu();
-    final Product product = Fixture.fixtureProduct();
-    final MenuGroup menuGroup = Fixture.fixtureMenuGroup();
+    final Menu menu = FixtureMenu.fixtureMenu();
+    final Product product = FixtureProduct.fixtureProduct();
+    final MenuGroup menuGroup = FixtureMenu.fixtureMenuGroup();
 
     given(menuGroupRepository.findById(any())).willReturn(Optional.of(menuGroup));
     given(productRepository.findAllByIdIn(any())).willReturn(List.of(product));
 
-    final Product failProduct = Fixture.fixtureProduct();
+    final Product failProduct = FixtureProduct.fixtureProduct();
     failProduct.setId(UUID.randomUUID());
     given(productRepository.findById(any())).willReturn(null);
 
@@ -122,9 +121,9 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴 금액이 상품 금액을 수량 만큼 합산한 금액 보다 클 수 없다.")
   void case7() {
-    final Menu menu = Fixture.fixtureMenu();
-    final Product product = Fixture.fixtureProduct();
-    final MenuGroup menuGroup = Fixture.fixtureMenuGroup();
+    final Menu menu = FixtureMenu.fixtureMenu();
+    final Product product = FixtureProduct.fixtureProduct();
+    final MenuGroup menuGroup = FixtureMenu.fixtureMenuGroup();
 
     menu.setPrice(BigDecimal.valueOf(30_000_000L));
 
@@ -140,9 +139,9 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴를 등록하면 해당 상품이 공개된다.")
   void case8() {
-    final Menu menu = Fixture.fixtureMenu();
-    final Product product = Fixture.fixtureProduct();
-    final MenuGroup menuGroup = Fixture.fixtureMenuGroup();
+    final Menu menu = FixtureMenu.fixtureMenu();
+    final Product product = FixtureProduct.fixtureProduct();
+    final MenuGroup menuGroup = FixtureMenu.fixtureMenuGroup();
 
     given(menuGroupRepository.findById(any())).willReturn(Optional.of(menuGroup));
     given(productRepository.findAllByIdIn(any())).willReturn(List.of(product));
@@ -157,8 +156,8 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴의 금액을 변경하기 위해 해당 메뉴가 존재해야 하고, 변경할 금액을 입력해야 한다.")
   void case9() {
-    final Menu menu = Fixture.fixtureMenu();
-    final Menu changeMenuPrice = Fixture.fixtureMenu();
+    final Menu menu = FixtureMenu.fixtureMenu();
+    final Menu changeMenuPrice = FixtureMenu.fixtureMenu();
 
     given(menuRepository.findById(any())).willReturn(Optional.of(menu));
 
@@ -172,8 +171,8 @@ class MenuServiceTest {
   @ValueSource(longs = {-1})
   @DisplayName("메뉴의 금액을 변경하기 위해 금액을 입력하지 않았거나 또는 0원 밑으로 입력할 수 없다.")
   void case10(Long value) {
-    final Menu menu = Fixture.fixtureMenu();
-    final Menu changeMenuPrice = Fixture.fixtureMenu();
+    final Menu menu = FixtureMenu.fixtureMenu();
+    final Menu changeMenuPrice = FixtureMenu.fixtureMenu();
 
     changeMenuPrice.setPrice(value == null ? null : BigDecimal.valueOf(value));
     Assertions.assertThatIllegalArgumentException()
@@ -183,8 +182,8 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴의 금액을 변경하기 위해 변경할 금액이 상품 금액을 수량 만큼 합산한 금액 보다 클 수 없다.")
   void case11() {
-    final Menu menu = Fixture.fixtureMenu();
-    final Menu changeMenuPrice = Fixture.fixtureMenu();
+    final Menu menu = FixtureMenu.fixtureMenu();
+    final Menu changeMenuPrice = FixtureMenu.fixtureMenu();
 
     given(menuRepository.findById(any())).willReturn(Optional.of(menu));
 
@@ -196,7 +195,7 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴를 노출하기 위해 메뉴가 존재해야한다.")
   void case12() {
-    final Menu menu = Fixture.fixtureMenu();
+    final Menu menu = FixtureMenu.fixtureMenu();
 
     menu.setDisplayed(false);
     given(menuRepository.findById(any())).willReturn(Optional.of(menu));
@@ -208,7 +207,7 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴를 노출하기 위해 메뉴 금액이 상품 금액을 수량 만큼 합산한 금액 보다 클 수 없다.")
   void case13() {
-    final Menu menu = Fixture.fixtureMenu();
+    final Menu menu = FixtureMenu.fixtureMenu();
 
     menu.setDisplayed(false);
     menu.setPrice(BigDecimal.valueOf(30_000L));
@@ -221,7 +220,7 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴를 숨기기 위해 메뉴가 존재해야 한다.")
   void case14() {
-    final Menu menu = Fixture.fixtureMenu();
+    final Menu menu = FixtureMenu.fixtureMenu();
 
     menu.setDisplayed(true);
     given(menuRepository.findById(any())).willReturn(Optional.of(menu));
@@ -233,7 +232,7 @@ class MenuServiceTest {
   @Test
   @DisplayName("메뉴를 전체 조회할 수 있다.")
   void case15() {
-    final Menu menu = Fixture.fixtureMenu();
+    final Menu menu = FixtureMenu.fixtureMenu();
     given(menuRepository.findAll()).willReturn(List.of(menu));
 
     final List<Menu> all = menuService.findAll();
