+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,9 +1,8 @@
 package kitchenpos.application;
 
 import kitchenpos.domain.*;
-import kitchenpos.fixtures.Fixture;
+import kitchenpos.fixtures.FixtureOrder;
 import kitchenpos.infra.KitchenridersClient;
-import org.assertj.core.api.Assert;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,7 +10,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.BDDMockito;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.math.BigDecimal;
@@ -19,7 +17,6 @@ import java.util.List;
 import java.util.Optional;
 
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
 
 @ExtendWith(SpringExtension.class)
 class OrderServiceTest {
@@ -35,7 +32,7 @@ class OrderServiceTest {
             "주문을 하기 위해 주문 정보에 메뉴가 한 개 이상 존재해야 한다.")
     @Test
     void case2() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final Menu menu = order.getOrderLineItems().get(0).getMenu();
         final OrderTable orderTable = order.getOrderTable();
 
@@ -52,7 +49,7 @@ class OrderServiceTest {
     @DisplayName("주문을 하기 위해 해당 메뉴의 상태가 노출돼야 한다.")
     @Test
     void case5() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final Menu menu = order.getOrderLineItems().get(0).getMenu();
         final OrderTable orderTable = order.getOrderTable();
 
@@ -69,7 +66,7 @@ class OrderServiceTest {
     @DisplayName("주문을 하기 위해 주문 정보에 금액과 메뉴 금액이 같아야 한다.")
     @Test
     void case6() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final Menu menu = order.getOrderLineItems().get(0).getMenu();
         final OrderTable orderTable = order.getOrderTable();
 
@@ -87,7 +84,7 @@ class OrderServiceTest {
     @DisplayName("주문 타입이 배달(DELIVERY)일 경우 주문이 들어와 있어야 하며, 주소지를 입력해야 한다.")
     @Test
     void case7() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final Menu menu = order.getOrderLineItems().get(0).getMenu();
         final OrderTable orderTable = order.getOrderTable();
 
@@ -105,7 +102,7 @@ class OrderServiceTest {
     @DisplayName("주문 타입이 먹고가기(EAT_IN)일 경우 주문테이블에 고객이 앉아있어야 하며 주문 수량은 1개 이상이어야 한다.")
     @Test
     void case8() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final Menu menu = order.getOrderLineItems().get(0).getMenu();
         final OrderTable orderTable = order.getOrderTable();
 
@@ -126,7 +123,7 @@ class OrderServiceTest {
             "주문을 접수(ACCEPTED)하기 위해 주문이 있어야 하고, 주문 타입이 대기중(WAITING)이여야 한다.")
     @Test
     void case9() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final Menu menu = order.getOrderLineItems().get(0).getMenu();
         final OrderTable orderTable = order.getOrderTable();
 
@@ -146,7 +143,7 @@ class OrderServiceTest {
     @DisplayName("주문을 접수(ACCEPTED)하기 위해 주문 타입이 배달(DELIVERY)이면 주문 정보와 총 금액, 배달지 주소를 배달기사에게 전달한다.")
     @Test
     void case11() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final OrderTable orderTable = order.getOrderTable();
 
         orderTable.setOccupied(true);
@@ -163,7 +160,7 @@ class OrderServiceTest {
     @DisplayName("주문을 전달(SERVED)하기 위해 주문 정보가 있어야 하고, 주문 타입이 접수(ACCEPTED)이여야 한다.")
     @Test
     void case12() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final OrderTable orderTable = order.getOrderTable();
 
         orderTable.setOccupied(true);
@@ -180,7 +177,7 @@ class OrderServiceTest {
     @DisplayName("주문을 배달중(DELIVERING)하기 위해 주문 정보가 있어야 하고, 주문 타입이 배달(DELIVERY)하고 주문 상태가 전달(SERVED)여야 한다.")
     @Test
     void case13() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final OrderTable orderTable = order.getOrderTable();
 
         orderTable.setOccupied(true);
@@ -197,7 +194,7 @@ class OrderServiceTest {
     @DisplayName("주문을 배달(DELIVERED)하기 위해 주문 정보가 있어야 하고, 주문 타입이 배달중(DELIVERING)여야 한다.")
     @Test
     void case14() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final OrderTable orderTable = order.getOrderTable();
 
         orderTable.setOccupied(true);
@@ -214,7 +211,7 @@ class OrderServiceTest {
     @DisplayName("주문을 완료(COMPLETED)하기 위해 주문 정보가 있어야 하고, 주문 타입이 배달중(DELIVERING), 그리고 주문 상태는 배달(DELIVERED)이여야 한다.")
     @Test
     void case15() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final OrderTable orderTable = order.getOrderTable();
 
         orderTable.setOccupied(true);
@@ -231,7 +228,7 @@ class OrderServiceTest {
     @DisplayName("주문을 완료(COMPLETED)하기 위해 주문 정보가 있어야 하고, 주문 타입이 가져가기(TAKEOUT) 또는 먹고가기(EAT_IN), 그리고 주문 상태는 전달(SERVED)이여야 한다.")
     @Test
     void case16() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final OrderTable orderTable = order.getOrderTable();
 
         orderTable.setOccupied(true);
@@ -247,7 +244,7 @@ class OrderServiceTest {
     @DisplayName("주문을 완료(COMPLETED)하게 되면, 고객이 음식을 먹고 갔기 때문에 주문테이블이 비어있어야 한다.")
     @Test
     void case17() {
-        final Order order = Fixture.fixtureOrder();
+        final Order order = FixtureOrder.fixtureOrder();
         final OrderTable orderTable = order.getOrderTable();
 
         orderTable.setOccupied(true);
@@ -266,7 +263,7 @@ class OrderServiceTest {
     @DisplayName("주문을 전체 조회 할 수 있다.")
     @Test
     void case18() {
-        final List<Order> orders = List.of(Fixture.fixtureOrder());
+        final List<Order> orders = List.of(FixtureOrder.fixtureOrder());
         BDDMockito.given(orderRepository.findAll()).willReturn(orders);
     }
