+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -12,7 +12,9 @@ import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuRepository;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.ProductRepository;
-import kitchenpos.fixtures.Fixture;
+import kitchenpos.fixtures.FixtureMenu;
+import kitchenpos.fixtures.FixtureOrder;
+import kitchenpos.fixtures.FixtureProduct;
 import kitchenpos.infra.PurgomalumClient;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -34,10 +36,10 @@ class ProductServiceTest {
   @Test
   @DisplayName("상품을 등록하기 위해 이름과 금액을 입력해야 한다.")
   void test_case_1() {
-    final Product expected = Fixture.fixtureProduct();
-    given(productRepository.save(any())).willReturn(Fixture.fixtureProduct());
+    final Product expected = FixtureProduct.fixtureProduct();
+    given(productRepository.save(any())).willReturn(FixtureProduct.fixtureProduct());
 
-    final Product product = productService.create(Fixture.fixtureProduct());
+    final Product product = productService.create(FixtureProduct.fixtureProduct());
 
     Assertions.assertThat(product.getName()).isEqualTo(expected.getName());
     Assertions.assertThat(product.getPrice()).isEqualTo(expected.getPrice());
@@ -46,7 +48,7 @@ class ProductServiceTest {
   @Test
   @DisplayName("상품을 등록하기 위해 이름에 비속어 또는 욕설을 작성할 수 없다.")
   void test_case_2() {
-    final Product product = Fixture.fixtureProduct();
+    final Product product = FixtureProduct.fixtureProduct();
 
     Mockito.when(purgomalumClient.containsProfanity(product.getName())).thenReturn(true);
 
@@ -60,9 +62,9 @@ class ProductServiceTest {
   @Test
   @DisplayName("상품의 금액을 수정")
   void test_case_3() {
-    final Product expected = Fixture.fixtureProduct();
+    final Product expected = FixtureProduct.fixtureProduct();
     given(productRepository.findById(any())).willReturn(Optional.of(expected));
-    given(menuRepository.findAllByProductId(any())).willReturn(List.of(Fixture.fixtureMenu()));
+    given(menuRepository.findAllByProductId(any())).willReturn(List.of(FixtureMenu.fixtureMenu()));
 
     expected.setPrice(BigDecimal.valueOf(10000L));
     final Product actual = productService.changePrice(UUID.randomUUID(), expected);
@@ -73,7 +75,7 @@ class ProductServiceTest {
   @Test
   @DisplayName("상품의 금액을 수정하기 위해 상품이 존재해야 하고, 금액을 입력하지 않았거나 또는 0원 밑으로 입력할 수 없다.")
   void test_case_4() {
-    final Product expected = Fixture.fixtureProduct();
+    final Product expected = FixtureProduct.fixtureProduct();
 
     expected.setPrice(BigDecimal.valueOf(-1));
     Assertions.assertThatIllegalArgumentException()
@@ -83,13 +85,14 @@ class ProductServiceTest {
   @Test
   @DisplayName("상품의 금액을 수정하면 해당 상품이 등록된 모든 메뉴에 영향이 간다.")
   void test_case_5() {
-    final Product expected = Fixture.fixtureProduct();
-    final List<Menu> menus = List.of(Fixture.fixtureMenu());
+    final List<Menu> menus = List.of(FixtureMenu.fixtureMenu());
+    final Product expected = menus.getFirst().getMenuProducts().getFirst().getProduct();
+    final Product request = FixtureProduct.fixtureProduct();
     given(productRepository.findById(any())).willReturn(Optional.of(expected));
     given(menuRepository.findAllByProductId(any())).willReturn(menus);
 
-    expected.setPrice(BigDecimal.valueOf(10L));
-    productService.changePrice(expected.getId(), expected);
+    request.setPrice(BigDecimal.valueOf(10L));
+    productService.changePrice(expected.getId(), request);
 
     Assertions.assertThat(menus.getFirst().isDisplayed()).isFalse();
   }
@@ -97,7 +100,7 @@ class ProductServiceTest {
   @Test
   @DisplayName("상품 전체를 조회할 수 있다.")
   void test_case_6() {
-    final List<Product> expected = List.of(Fixture.fixtureProduct());
+    final List<Product> expected = List.of(FixtureProduct.fixtureProduct());
     given(productRepository.findAll()).willReturn(expected);
 
     final List<Product> actual = productService.findAll();
